feat(files): merge components
