+++ src/components/DataDisplay/FileViewer/FileViewer.tsx
@@ -1,48 +1,123 @@
-import { Box } from "@mui/material";
+import { Box, SxProps, Theme } from "@mui/material";
+import { useCallback, useEffect, useRef, useState } from "react";
+import notFoundImage from "@/assets/img/not-found-img.svg";
 import { Lightbox } from "@/main";
 
-interface FullSizeViewProps {
-  src: string;
+interface FileViewerPros {
+  srcThumbnail: string;
+  srcFileViewer?: string;
   fileName?: string;
-  open: boolean;
-  onClose?(): void;
-  width?: number | string;
-  height?: number | string;
+  thumbWidth?: number | string;
+  thumbHeight?: number | string;
+  fileViewWidth?: number | string;
+  fileViewHeight?: number | string;
+  sx?: SxProps;
 }
 
 const styles = {
-  viewFile: {
+  container: {
+    ":hover": {
+      opacity: 0.8,
+    },
+    alignSelf: "center",
+    backgroundColor: ({ palette }: Theme) => palette.background.paper,
+    boxShadow: 4,
+    overflow: "hidden",
+    position: "relative",
+  },
+  viewClickFile: {
     border: 0,
     height: "100vh",
     width: "50vw",
   },
-  viewImage: {
+  viewClickImage: {
     border: 0,
     height: "100%",
     width: "100%",
   },
+  viewerFile: {
+    border: 0,
+    objectFit: "cover",
+    pointerEvents: "none",
+  },
+  viewerImage: {
+    border: 0,
+    objectFit: "cover",
+  },
 };
 
-const FileViewer = ({ src, fileName, onClose, open, width = 100, height = 80 }: FullSizeViewProps) => {
-  const isImage = /\.(png|jpe?g|gif|bmp|webp|avif|tiff?|svg)$/i.test(src.toLowerCase()) && !src.startsWith("blob:");
-  const isPdf = /\.pdf$/i.test(src.toLowerCase());
+const FileViewer = ({
+  srcThumbnail,
+  srcFileViewer,
+  thumbWidth,
+  thumbHeight,
+  fileName,
+  fileViewWidth,
+  fileViewHeight,
+  sx,
+}: FileViewerPros) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const [isError, setIsError] = useState(false);
+  const isImage = !srcThumbnail?.endsWith(".pdf") && !srcThumbnail.startsWith("blob:");
+  const iframeRef = useRef<HTMLObjectElement>(null);
+  const data = isError ? notFoundImage : srcThumbnail;
+  const isPdf = srcFileViewer && /\.pdf$/i.test(srcFileViewer.toLowerCase());
 
-  if (!isPdf && !isImage) {
-    return null;
-  }
+  const handleToggleOpen = () => {
+    setIsOpen(!isOpen);
+  };
+
+  const errorHandle = useCallback(() => {
+    setIsError(true);
+  }, []);
+
+  // Handle error
+  useEffect(() => {
+    const object = iframeRef.current;
+    object?.addEventListener("error", errorHandle);
+
+    return () => {
+      object?.removeEventListener("error", errorHandle);
+    };
+  }, [errorHandle, iframeRef]);
 
   return (
-    <Lightbox open={open} onClose={onClose} src={src} title={fileName}>
-      {isImage ? null : (
+    <>
+      <Box
+        data-test="fileViewer"
+        onClick={handleToggleOpen}
+        sx={{
+          ...styles.container,
+          cursor: "pointer",
+          display: "block",
+          flexBasis: thumbWidth,
+          flexShrink: 0,
+          thumbHeight,
+          thumbWidth,
+          ...sx,
+        }}
+      >
         <Box
           component={isImage ? "img" : "iframe"}
-          width={width}
-          height={height}
-          src={src}
-          sx={isImage ? styles.viewImage : styles.viewFile}
+          key={data}
+          overflow="hidden"
+          height={isImage ? "100%" : "auto"}
+          width="100%"
+          src={data}
+          ref={iframeRef}
+          sx={isImage ? styles.viewerImage : styles.viewerFile}
         />
+      </Box>
+      {srcFileViewer && (
+        <Lightbox open={isOpen} onClose={handleToggleOpen} src={srcFileViewer} title={fileName}>
+          {isPdf ? (
+            <Box component="iframe" width={fileViewWidth} height={fileViewHeight} src={srcFileViewer} sx={styles.viewClickFile} />
+          ) : (
+            <Box component="img" width={fileViewWidth} height={fileViewHeight} src={srcFileViewer} sx={styles.viewClickImage} />
+          )}
+        </Lightbox>
       )}
-    </Lightbox>
+    </>
   );
 };
