+++ src/components/DataDisplay/FileViewer/stories/FileViewer.stories.tsx
@@ -1,35 +1,33 @@
-import { Stack, Button } from "@mui/material";
+import { Stack } from "@mui/material";
 import type { Meta, StoryFn } from "@storybook/react";
-import React, { useState } from "react";
 import FileViewer from "@/components/DataDisplay/FileViewer";
 
-const Template: StoryFn<typeof FileViewer> = (args) => {
-  const [isOpen, setIsOpen] = useState(false);
-  const handleOpen = () => setIsOpen(true);
-  const handleClose = () => setIsOpen(false);
+const Template: StoryFn<typeof FileViewer> = (args) => (
+  <Stack height={200}>
+    <FileViewer {...args} />
+  </Stack>
+);
 
-  return (
-    <Stack height="100%" spacing={2}>
-      <Button variant="contained" onClick={handleOpen}>
-        open PDF
-      </Button>
-
-      <FileViewer {...args} open={isOpen} onClose={handleClose} />
-
-      {isOpen && (
-        <Button variant="contained" onClick={handleClose}>
-          close PDF
-        </Button>
-      )}
-    </Stack>
-  );
+export const PdfThumbnailViewer = Template.bind({});
+PdfThumbnailViewer.args = {
+  srcThumbnail: "https://pousses.fr/sites/default/files/2019-08/pdf_test_1.pdf",
 };
 
 export const PdfFileViewer = Template.bind({});
 PdfFileViewer.args = {
-  onClose: () => {},
-  open: false,
-  src: "https://pousses.fr/sites/default/files/2019-08/pdf_test_1.pdf",
+  srcFileViewer: "https://pousses.fr/sites/default/files/2019-08/pdf_test_1.pdf",
+  srcThumbnail: "https://pousses.fr/sites/default/files/2019-08/pdf_test_1.pdf",
+};
+
+export const ImageThumbnailViewer = Template.bind({});
+ImageThumbnailViewer.args = {
+  srcThumbnail: "https://upload.wikimedia.org/wikipedia/commons/6/6a/PNG_Test.png",
+};
+
+export const ImageFileViewer = Template.bind({});
+ImageFileViewer.args = {
+  srcFileViewer: "https://upload.wikimedia.org/wikipedia/commons/6/6a/PNG_Test.png",
+  srcThumbnail: "https://upload.wikimedia.org/wikipedia/commons/6/6a/PNG_Test.png",
 };
 
 export default {
