--- src/components/DataDisplay/ThumbnailFileViewer/index.ts
@@ -1,4 +1,0 @@
-import ThumbnailFileViewer from "./ThumbnailFileViewer";
-
-export default ThumbnailFileViewer;
-export * from "./ThumbnailFileViewer";
