--- src/components/DataDisplay/ThumbnailViewer/index.ts
@@ -1,4 +1,0 @@
-import ThumbnailViewer from "./ThumbnailViewer";
-
-export default ThumbnailViewer;
-export * from "./ThumbnailViewer";
